[BE] Enable ZeRO test on windows
Enables the ZeRO tests to run on windows. Closes
https://github.com/pytorch/pytorch/issues/63086.

Backend == NCCL was used as a proxy to see if we were running under CUDA, but Windows GPU tests uses Gloo. In this case use Gloo on GPU.

For some reason these tests don't seem to test Gloo on GPU with ZeRO in general (picks NCCL backend when GPU is available), so kept that behavior for now.

Differential Revision: [D31071181](https://our.internmc.facebook.com/intern/diff/D31071181/)

ghstack-source-id: 138559877
Pull Request resolved: https://github.com/pytorch/pytorch/pull/65385